refactor(scroll-container): move component
