+++ src/bricks/index.ts
@@ -15,3 +15,4 @@ export {
 } from './layout/layout';
 export { Input } from './input/input';
 export { Modal, useModal } from './modal/modal';
+export { ScrollContainer } from './scroll-container/scroll-container';
